test: split app tests with shared helpers
